+++ ppm_image.h
@@ -17,7 +17,7 @@ typedef struct {
 
 // Structure representing the image
 typedef struct {
-    int height, width;
+    size_t height, width;
     size_t totalPixels; // biggest type to store the unknown and possibly big total number
     pixel_t *data; // dynamic array of pixels
 } ppm_image_t;
@@ -47,8 +47,8 @@ int ppm_image_t_height(const ppm_image_t *ppmImage);
 size_t ppm_image_t_totalPixels(const ppm_image_t *ppmImage);
 size_t ppm_black_pixels(const ppm_image_t *img);
 size_t ppm_black_pixels_flex(const ppm_image_t *img, int accuracy);
-void* ppm_black_pixels_T1 (void* arg);
-void* ppm_black_pixels_T2 (void* arg);
+void* ppm_black_pixels_T1(void* arg);
+void* ppm_black_pixels_T2(void* arg);
 pixel_t pixel_new(uint8_t red, uint8_t green, uint8_t blue);
 pixel_t *ppm_image_t_data(const ppm_image_t *ppmImage);
 pixel_t ppm_pixel(const ppm_image_t *img, size_t x, size_t y);
